Add files via upload
This script takes a file with all the suid binaries running in the target system and checks it against the list of SUID binaries listed on gtfobins (https://gtfobins.github.io/#+suid) to look for SUID binaries on the target system which can be used escalate our privileges in unix/linux environment. 

The purpose of creating this script is to save time by automatically finding the exploitable SUID binaries without having to manually go through them in the target system.